requirements.txt
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.8
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
cycler==0.11.0
decorator==5.1.1
docx==0.2.4
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fonttools==4.38.0
geopandas==0.10.2
h11==0.14.0
idna==3.10
importlib-metadata==6.7.0
ipython==7.34.0
jedi==0.19.1
kiwisolver==1.4.5
lxml==5.3.0
matplotlib-inline==0.1.6
numpy==1.26.4
openpyxl==3.1.3
outcome==1.3.0.post0
packaging==24.0
pandas==1.3.5
parso==0.8.4
patsy==0.5.6
pickleshare==0.7.5
prompt_toolkit==3.0.47

requests==2.31.0

shapely==2.0.7

typing_extensions==4.7.1

